requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
requests>=2.31.0
jinja2>=3.1.0
sentence-transformers>=2.7.0
numpy>=1.26.0
faiss-cpu>=1.12.0
pypdf>=6.1.0
python-dotenv>=1.0.1
